Make MongoSession package private
See gh-2170
