+++ spring-session-data-mongodb/src/test/java/org/springframework/session/data/mongo/config/annotation/web/reactive/ReactiveMongoWebSessionConfigurationTest.java
@@ -34,12 +34,12 @@ import org.springframework.data.mongodb.core.ReactiveMongoOperations;
 import org.springframework.data.mongodb.core.index.IndexOperations;
 import org.springframework.session.IndexResolver;
 import org.springframework.session.ReactiveSessionRepository;
+import org.springframework.session.Session;
 import org.springframework.session.config.ReactiveSessionRepositoryCustomizer;
 import org.springframework.session.config.annotation.web.server.EnableSpringWebSession;
 import org.springframework.session.data.mongo.AbstractMongoSessionConverter;
 import org.springframework.session.data.mongo.JacksonMongoSessionConverter;
 import org.springframework.session.data.mongo.JdkMongoSessionConverter;
-import org.springframework.session.data.mongo.MongoSession;
 import org.springframework.session.data.mongo.ReactiveMongoSessionRepository;
 import org.springframework.util.ReflectionUtils;
 import org.springframework.web.server.adapter.WebHttpHandlerBuilder;
@@ -189,7 +189,7 @@ public class ReactiveMongoWebSessionConfigurationTest {
 		this.context.refresh();
 
 		ReactiveMongoSessionRepository repository = this.context.getBean(ReactiveMongoSessionRepository.class);
-		IndexResolver<MongoSession> indexResolver = this.context.getBean(IndexResolver.class);
+		IndexResolver<Session> indexResolver = this.context.getBean(IndexResolver.class);
 
 		assertThat(repository).isNotNull();
 		assertThat(indexResolver).isNotNull();
@@ -205,7 +205,7 @@ public class ReactiveMongoWebSessionConfigurationTest {
 		this.context.refresh();
 
 		ReactiveMongoSessionRepository repository = this.context.getBean(ReactiveMongoSessionRepository.class);
-		IndexResolver<MongoSession> indexResolver = this.context.getBean(IndexResolver.class);
+		IndexResolver<Session> indexResolver = this.context.getBean(IndexResolver.class);
 
 		assertThat(repository).isNotNull();
 		assertThat(indexResolver).isNotNull();
@@ -367,7 +367,7 @@ public class ReactiveMongoWebSessionConfigurationTest {
 
 		@Bean
 		@SuppressWarnings("unchecked")
-		IndexResolver<MongoSession> indexResolver() {
+		IndexResolver<Session> indexResolver() {
 			return mock(IndexResolver.class);
 		}
 
@@ -389,7 +389,7 @@ public class ReactiveMongoWebSessionConfigurationTest {
 
 		@Bean
 		@SuppressWarnings("unchecked")
-		IndexResolver<MongoSession> indexResolver() {
+		IndexResolver<Session> indexResolver() {
 			return mock(IndexResolver.class);
 		}
